- Fixed bug #60337 bcscale related crashed on 64bits platforms
git-svn-id: http://svn.php.net/repository/php/php-src/trunk@319546 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/bcmath/bcmath.c
@@ -560,7 +560,7 @@ PHP_FUNCTION(bcscale)
 		return;
 	}
 
-	BCG(bc_precision) = (new_scale < 0) ? 0 : new_scale;
+	BCG(bc_precision) = ((int)new_scale < 0) ? 0 : new_scale;
 
 	RETURN_TRUE;
 }
